+++ package.json
@@ -30,6 +30,7 @@
     "playwright": "1.8.1",
     "prettier": "^2.2.1",
     "ts-jest": "^26.4.4",
+    "ts-node": "^9.1.1",
     "typedoc": "^0.20.24",
     "typedoc-neo-theme": "^1.0.10",
     "typescript": "^4.1.2"
